+++ provision.sh
@@ -16,6 +16,7 @@ if [ ! -e /usr/local/bin/node ]; then
 fi
 
 sudo npm install -g coffee-script@1.7.1
+sudo npm install -g livescript@1.4.0
 sudo npm install -g less@1.7.4
 sudo npm install -g babel@4.3.0