+++ static_precompiler/compilers/__init__.py
@@ -4,3 +4,4 @@ from .babel import *  # noqa
 from .scss import *  # noqa
 from .less import *  # noqa
 from .stylus import *  # noqa
+from .livescript import *  # noqa